requirements.txt

pydantic>=2.4
yfinance>=0.2.40
pandas>=2.1
numpy>=1.26
pandas-ta>=0.3.14b
requests>=2.31
feedparser>=6.0
vaderSentiment>=3.3
rich>=13.7
python-dateutil>=2.9
fastapi>=0.111
uvicorn>=0.30
pytest>=8.0
